Major restructing; also bug fixes
Restructured it so GetQP/MS are treated as what they are: dialogs. Now
we open the MainWindow first, but in an invisible state, and call those
two as dialogs. This way we can avoid the use of global variables for
the results of these dialogs.

Also fixed the infamous mark scheme window hang - it turns out that if
you have an open web request (in our case the one before, from
downloading the question paper) attempting to download a file from the
same URL as the request will cause the program to hang at that point.